Update pre-commit and rename async because it will be reserved in py37
